+++ RevitAddin.DA.Tester/Services/DesignAutomationController.cs
@@ -48,6 +48,14 @@ namespace RevitAddin.DA.Tester.Services
             Console.WriteLine($"Shape Location:\t{typeof(ricaun.Revit.DB.Shape.Colors).Assembly.Location}");
             Console.WriteLine("----------------------------------------");
 
+            application.DocumentCreated += (s, e) => {
+                Console.WriteLine($"DocumentCreated:\t{e.Document.Title}");
+                Console.WriteLine($"DocumentCreated.AddInName: \t{application.ActiveAddInId?.GetAddInName()}");
+                Console.WriteLine("----------------------------------------");
+            };
+
+            application.NewProjectDocument(UnitSystem.Metric);
+
             return true;
         }
     }